Split test class in two classes
